requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0

# AI/ML Framework - Compatible versions
langchain==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langchain-core==0.3.0
langchain-community==0.3.0
openai==1.54.0
anthropic==0.39.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# API & Web
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
websockets==12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
loguru==0.7.2

# Validation & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring & Alerts
schedule==1.2.0

# Fuzzy matching for edge cases
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Optional: N8N Integration
requests==2.31.0
